+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.6.19
 
+* New `Completion` subcommand to generate shell completions.
 * Fix: An issue with not reserving enough memories for the largest possible string if the octet length reported by the driver is to small. Now calculation is based on column sized.
 
 ## 0.6.18